+++ how-to/customize-home-templates/client/src/shapes.ts
@@ -41,12 +41,17 @@ export interface ISource {
 	 * @param query The query to search for.
 	 * @param filters The filters to apply.
 	 * @param lastResponse The last search response used for updating existing results.
+	 * @param options Options for the get search results.
+	 * @param options.isSuggestion Is the query a suggestion.
 	 * @returns The list of results and new filters.
 	 */
 	getSearchResults(
 		query: string,
 		filters: CLIFilter[],
-		lastResponse: HomeSearchListenerResponse
+		lastResponse: HomeSearchListenerResponse,
+		options: {
+			isSuggestion: boolean;
+		}
 	): Promise<HomeSearchResponse>;
 
 	/**