+++ how-to/register-with-dock/client/webpack.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-/*  */
+
 module.exports = [
 	{
 		entry: './client/src/provider.ts',
@@ -18,6 +18,26 @@ module.exports = [
 		},
 		output: {
 			filename: 'provider.bundle.js',
+			path: path.resolve(__dirname, '..', 'public', 'js')
+		}
+	},
+	{
+		entry: './client/src/custom-menu.ts',
+		devtool: 'inline-source-map',
+		module: {
+			rules: [
+				{
+					test: /\.tsx?$/,
+					use: 'ts-loader',
+					exclude: /node_modules/
+				}
+			]
+		},
+		resolve: {
+			extensions: ['.tsx', '.ts', '.js']
+		},
+		output: {
+			filename: 'custom-menu.bundle.js',
 			path: path.resolve(__dirname, '..', 'public', 'js')
 		}
 	}