+++ how-to/workspace-platform-starter/client/src/framework/shapes/module-shapes.ts
@@ -1,9 +1,8 @@
 import type OpenFin from "@openfin/core";
-import type { InteropClient } from "@openfin/core/src/api/interop";
 import type { BrowserWindowModule, CustomPaletteSet, Page } from "@openfin/workspace-platform";
 import type { PlatformApp } from "./app-shapes";
 import type { LifecycleEvents, LifecycleHandler } from "./lifecycle-shapes";
-import type { LoggerCreator } from "./logger-shapes";
+import type { Logger, LoggerCreator } from "./logger-shapes";
 import type { ColorSchemeMode } from "./theme-shapes";
 import type { VersionInfo } from "./version-shapes";
 
@@ -119,7 +118,7 @@ export interface ModuleHelpers {
 	 * cache it and use it once the application is bootstrapped and ready.
 	 * @returns The interop client.
 	 */
-	getInteropClient?(): Promise<InteropClient | undefined>;
+	getInteropClient?(): Promise<OpenFin.InteropClient | undefined>;
 
 	/**
 	 * If available, this function lets you request the launch of an application that is available to this platform and
@@ -132,10 +131,22 @@ export interface ModuleHelpers {
 	/**
 	 * Launch a page in the workspace.
 	 * @param page The page to launch.
-	 * @param bounds The optional bounds for the page.
+	 * @param options The options for the launch.
+	 * @param options.bounds The optional bounds for the page.
+	 * @param options.targetWindowIdentity The optional target window for the page.
+	 * @param options.createCopyIfExists Create a copy of the page if it exists.
+	 * @param logger Log output from the operation.
 	 * @returns The window created.
 	 */
-	launchPage?(page: Page, bounds?: OpenFin.Bounds): Promise<BrowserWindowModule>;
+	launchPage?(
+		page: Page,
+		options?: {
+			bounds?: OpenFin.Bounds;
+			targetWindowIdentity?: OpenFin.Identity;
+			createCopyIfExists?: boolean;
+		},
+		logger?: Logger
+	): Promise<BrowserWindowModule>;
 
 	/**
 	 * Subscribe to lifecycle events.