+++ how-to/workspace-platform-starter/client/src/framework/shapes/setting-shapes.ts
@@ -17,6 +17,7 @@ import type { LowCodeIntegrationProviderOptions } from "./low-code-integration-s
 import type { MenusProviderOptions } from "./menu-shapes";
 import type { NotificationProviderOptions } from "./notification-shapes";
 import type { PlatformProviderOptions } from "./platform-shapes";
+import type { SplashScreenProviderOptions } from "./splash-shapes";
 import type { StorefrontProviderOptions } from "./store-shapes";
 import type { ThemeProviderOptions } from "./theme-shapes";
 import type { VersionProviderOptions } from "./version-shapes";
@@ -139,4 +140,9 @@ export interface CustomSettings {
 	 * The menus provider options.
 	 */
 	menusProvider?: MenusProviderOptions;
+
+	/**
+	 * The splash screen provider options.
+	 */
+	splashScreenProvider?: SplashScreenProviderOptions;
 }