+++ how-to/workspace-platform-starter/client/src/modules/composite/about/integration.ts
@@ -129,6 +129,7 @@ export class AboutProvider implements IntegrationModule<AboutProviderSettings> {
 	 * @param options Options for the search query.
 	 * @param options.queryMinLength The minimum length before a query is actioned.
 	 * @param options.queryAgainst The fields in the data to query against.
+	 * @param options.isSuggestion Is the query from a suggestion.
 	 * @returns The list of results and new filters.
 	 */
 	public async getSearchResults(
@@ -138,6 +139,7 @@ export class AboutProvider implements IntegrationModule<AboutProviderSettings> {
 		options: {
 			queryMinLength: number;
 			queryAgainst: string[];
+			isSuggestion?: boolean;
 		}
 	): Promise<HomeSearchResponse> {
 		if (query.length < 2 || !AboutProvider._ABOUT_COMMAND.startsWith(query)) {