+++ how-to/workspace-platform-starter/client/src/modules/integrations/pages/integration.ts
@@ -5,10 +5,9 @@ import type {
 	HomeDispatchedSearchResult,
 	HomeSearchListenerResponse,
 	HomeSearchResponse,
-	HomeSearchResult,
-	Page
+	HomeSearchResult
 } from "@openfin/workspace";
-import type { WorkspacePlatformModule } from "@openfin/workspace-platform";
+import type { Page, WorkspacePlatformModule } from "@openfin/workspace-platform";
 import type {
 	IntegrationHelpers,
 	IntegrationModule,
@@ -154,6 +153,7 @@ export class PagesProvider implements IntegrationModule<PagesSettings> {
 	 * @param options Options for the search query.
 	 * @param options.queryMinLength The minimum length before a query is actioned.
 	 * @param options.queryAgainst The fields in the data to query against.
+	 * @param options.isSuggestion Is the query from a suggestion.
 	 * @returns The list of results and new filters.
 	 */
 	public async getSearchResults(
@@ -163,6 +163,7 @@ export class PagesProvider implements IntegrationModule<PagesSettings> {
 		options: {
 			queryMinLength: number;
 			queryAgainst: string[];
+			isSuggestion?: boolean;
 		}
 	): Promise<HomeSearchResponse> {
 		let pageResults: HomeSearchResult[] = [];
@@ -210,7 +211,7 @@ export class PagesProvider implements IntegrationModule<PagesSettings> {
 					if (this._integrationHelpers?.getPlatform && this._integrationHelpers?.launchPage) {
 						const platform = this._integrationHelpers.getPlatform();
 						const pageToLaunch = await platform.Storage.getPage(data.pageId);
-						await this._integrationHelpers.launchPage(pageToLaunch);
+						await this._integrationHelpers.launchPage(pageToLaunch, undefined, this._logger);
 					}
 				} else if (result.action.name === PagesProvider._ACTION_DELETE_PAGE) {
 					if (this._integrationHelpers?.getPlatform) {