+++ how-to/workspace-platform-starter/client/src/modules/integrations/workspaces/integration.ts
@@ -190,6 +190,7 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 	 * @param options Options for the search query.
 	 * @param options.queryMinLength The minimum length before a query is actioned.
 	 * @param options.queryAgainst The fields in the data to query against.
+	 * @param options.isSuggestion Is the query from a suggestion.
 	 * @returns The list of results and new filters.
 	 */
 	public async getSearchResults(
@@ -199,6 +200,7 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 		options: {
 			queryMinLength: number;
 			queryAgainst: string[];
+			isSuggestion?: boolean;
 		}
 	): Promise<HomeSearchResponse> {
 		if (this._integrationHelpers?.getPlatform && this._settings) {